First set of slides
I'm in the process of creating beamer slides for Piketty's book.

To begin with, I am going to replicate the powerpoint slides Piketty
used at a Sao Paulo presentation in December 2014 (because they are the
latest slides I found). I have just completed the first of 3 pdfs
(short version). In the process, I have refactored some of the code.
However, I would welcome suggestions about best practice so that my
code may be subsequently merged into the main project.

1. The data is defined in the preamble of the ".Rnw" document, which is
the master file for the 'Slides' project. I have named dataframes used
to produce the book figures according to the figure number in the book,
e.g. df15.1 for the dataframe used to produce figure 15.1. Sometimes
the data comes from a worksheet labeled TS15.1 in the original excel
files, but sometimes it comes from a different sheet, because the same
sheet might be used to produce several figures and/or because something
is unclear in the naming scheme used by Piketty in his worksheets. For
this reason, when I first load the data I name the dataframe ts15.1 and
later change it to df15.1 or df15.2, etc. as the case may be.

2. I have not used too many special functions to produce the plots
because I think that to be most useful it should be reasonably easy for
anyone to cut/paste the code for a single figure. There are 2
exceptions to this: a shapePalette and a colorPalette which I define in
the preamble. I have also defined a black and white scheme to produce
figures quite close to the original book style.

3. In this incarnation of my code-in-progress, I am producing the
graphics on the fly as the beamer is produced. In a later refactoring,
it may be desirable to produce all the graphics by Chapter, as the
current repository does, and simply use includegraphics in the beamer
presentation, pointing to the location of the figures in the book
repository.

4. Ultimately, it may be better to store each frame in a separate file
and to include them as the need arises. This would make it easier to
produce short/long presentations.

5. Once all the book figures have been produced in a consistent way and
dataframes named in an intuitive way, e.g. df15.1, I would save the
dataframes in a .rda file. And perhaps create a 'piketty' package with
the stored data.