Split data and slides into 2 separate '.Rnw' files
1. ‘slidesFigures.Rnw’ is the knitr/beamer document with all the Figure
slides (but no tables)

2. ‘dataFiguresTables.Rnw’ is the knitr/R document with all the data
creation for both the excel sheets and the Figures data frames (about
half of Figures done)

3. ‘slidesPikettyAll.tex’ contains slides copied from Piketty’s
presentations, formated to beamer style. It does not include the
Figures but does contain Tables (tables will be moved to a separate
document if I get to reproduce enough of them)

4. ‘slidesPikettySaoPaulo2014.tex’  lists the frames (by their labels)
that are to be included in the Sao Paulo 2014 presentation.

5. ‘slidesCurrent.tex’ and ‘slidesTesting.tex’ are for debugging
purposes and will be removed eventually.